requirements.txt
# for dataset download
kagglehub

# core
pytorch
numpy
opencv-python